requirements-dev.txt
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto[s3]>=5.0.0
responses>=0.25.0
freezegun>=1.4.0

# Linting & Formatting
ruff>=0.2.0
mypy>=1.8.0
boto3-stubs[s3]>=1.34.0

# Type stubs
types-requests>=2.31.0
